requirements.txt
PyYAML==6.0
tensorflow==2.13.0
Pillow==9.5.0
numpy==1.23.4
opencv-python==4.8.0.76
scikit-image==0.19.3
transformers==4.27.4